requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asttokens==3.0.0
attrs==23.2.0
boto3==1.37.13
botocore==1.37.13
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
dill==0.3.9
distro==1.9.0
docker==7.1.0
executing==2.2.0
fastapi==0.115.11
google-pasta==0.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==6.11.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mock==4.0.3
multiprocess==0.70.17
nest-asyncio==1.6.0
numpy==1.26.4
omegaconf==2.3.0
openai==1.66.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathos==0.3.3
pexpect==4.9.0
platformdirs==4.3.6
pox==0.3.5
ppft==1.7.6.9
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
s3transfer==0.11.4
sagemaker==2.242.0
sagemaker-core==1.0.25
schema==0.7.7
six==1.17.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.1
tblib==3.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
zipp==3.21.0
